% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_df}
\alias{task_df}
\title{Convert task as list to data frame}
\usage{
task_df(task)
}
\arguments{
\item{task}{as list}
}
\value{
a data frame
}
\description{
Some fields are renamed. Log fields are dicarded. Also add elapsed time fields.
Task name is splitted into two columns: workflow and task.
}
